Add heater input_channel None with id 0
The Lakeshore336 has a hardware option for selection a PID input_channel to be 'None' for heater outputs. This is currently not reflected in the driver and the LakeshoreModel336CurrentSource will fail to get or set the input_channel remotely if it was set to 'None' manually on the hardware.